wrl-dependencies.gv: Added source file for dependency graph
wrl-dependencies.gv is a hard coded graph of dependencies in the current
build. Note that there's no simple way to generate it automatically from
the source files as of now. A script should be made for this purpose.
The graph shows all the WRL files in the project with dependencies
marked. The shape of the arrow shows whether the dependency is Inline or
prototype.